Feature/ledger

+++ ledger/views.py
@@ -3,7 +3,6 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.generics import (
     get_object_or_404,
-    GenericAPIView,
 )
 from rest_framework import (
     status,
@@ -13,10 +12,15 @@ from rest_framework.permissions import IsAuthenticated
 from ledger.serializers import (
     RecordSummarySerializer,
     LedgerDetailSerializer,
+    RecordModifySerializer,
 )
 from drf_yasg.utils import swagger_auto_schema
 from ledger.models import Record
-from ledger.ledger_api_params import record_post_params
+from ledger.ledger_api_params import (
+    record_post_params,
+    record_modify_params,
+)
+
 from user.models import User
 # Create your views here.
 
@@ -31,54 +35,69 @@ class RecordListView(APIView):
     
     @swagger_auto_schema(request_body=record_post_params)
     def post(self, request):
-        serializer = LedgerDetailSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
-    
-class DetailAPIView(GenericAPIView):
+        user = User.objects.get(email=request.user)
+        record = Record.objects.create(user=user, **request.data)
+        serializer = LedgerDetailSerializer(record)
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+
+class DetailAPIView(APIView):
     permission_classes = [IsAuthenticated]
-    serializer_class = LedgerDetailSerializer
 
     def get(self, request, pk):
-        record = get_object_or_404(Record, id=pk)
-        serializer = self.serializer_class(record)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    
+        record = Record.obejcts.get(id=pk)
+        if record:
+            serializer = LedgerDetailSerializer(record)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+
+    @swagger_auto_schema(request_body=record_modify_params)
     def patch(self, request, pk):
-        record = get_object_or_404(Record, id=pk)
-        serializer = self.serializer_class(record)
-        serializer.update(record, request.data)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    
+        record = Record.objects.get(id=pk)
+        if record:
+            modifier_serializer = RecordModifySerializer(record)
+            modifier_serializer.update(record, request.data)
+            record = Record.objects.get(id=pk)
+            detail_serializer = LedgerDetailSerializer(record)
+            return Response(detail_serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+
     def delete(self, request, pk):
         record = get_object_or_404(Record, id=pk)
-        record.delete()
-        serializer = self.serializer_class(record)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+        if record:
+            record.delete()
+            serializer = LedgerDetailSerializer(record)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+
     
-class DeletedRecordListView(GenericAPIView):
+class DeletedRecordListView(APIView):
     permission_classes = [IsAuthenticated]
-    serializer_class = RecordSummarySerializer
 
     def get(self, request):
         user = User.objects.get(email=request.user)
-        records = Record.deleted_objects.all().filter(user=user)
-        serializer = self.serializer_class(records, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    
-class DeletedRecordDetailView(GenericAPIView):
+        if user:
+            records = Record.deleted_objects.all().filter(user=user)
+            serializer = RecordSummarySerializer(records, many=True)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+
+
+class DeletedRecordDetailView(APIView):
     permission_classes = [IsAuthenticated]
-    serializer_class = LedgerDetailSerializer
 
     def get(self, request, pk):
         record = Record.deleted_objects.get(id=pk)
-        serializer = self.serializer_class(record)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+        if record:
+            serializer = LedgerDetailSerializer(record)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
     
     def put(self, request, pk):
-        record=Record.deleted_objects.get(id=pk)
-        record.restore()
-        serializer = self.serializer_class(record)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+        record = Record.deleted_objects.get(id=pk)
+        if record:
+            record.restore()
+            serializer = LedgerDetailSerializer(record)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+
